+++ src/components/index.ts
@@ -1,5 +1,6 @@
 import Search from './Search';
 import ListItems from './ListItem';
 import Pagination from './Pagination';
+import ErrorBoundary from './Error';
 
-export { ListItems, Search, Pagination };
+export { ListItems, Search, Pagination, ErrorBoundary };
