+++ src/loaders/index.ts
@@ -1,0 +1,4 @@
+import fetchData from './fetchData';
+import initFetchData from './initFetchData';
+
+export { fetchData, initFetchData };
